test/topology: isolate topology smp params test
Move slow test for different smp parameters to its own file.
